+++ src/main/java/com/example/wiretransfers/DAOs/TransactionDAO.java
@@ -14,11 +14,11 @@ public interface TransactionDAO extends CrudRepository<Transaction, UUID> {
 
     public Transaction findByTransactionId(String transactionId);
 
-    public List<Transaction> findByReceiverAccountNumber(Long number);
+    public List<Transaction> findByReceiverAccountNumber(int number);
 
-    public List<Transaction> findBySenderAccountNumber(Long number);
+    public List<Transaction> findBySenderAccountNumber(int number);
 
-    public List<Transaction> findBySenderAccountNumberAndReceiverAccountNumber(Long sender, Long receiver);
+    public List<Transaction> findBySenderAccountNumberAndReceiverAccountNumber(int sender, int receiver);
 
     public List<Transaction> findByTimestampBetween(Date start, Date end);
